examples/python/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-ibm-api-connect==0.0.1-alpha.1703062432+5f7e0407.dirty
